Added Item charge module to the battery block Removed old item energy api to use forge energy instead

+++ src/main/kotlin/com/cout970/magneticraft/config/Config.kt
@@ -70,7 +70,7 @@ object Config {
     var solarPanelMaxProduction = 10.0
 
     @ConfigValue(category = CATEGORY_ENERGY, comment = "Item Battery Capacity")
-    var itemBatteryCapacity: Double = 250000.0
+    var itemBatteryCapacity: Int = 250000
 
     @ConfigValue(category = CATEGORY_ENERGY, comment = "Battery Block Capacity")
     var blockBatteryCapacity: Int = 1000000
